MH-236: fix module serializer to extend from searchable
